Moved Utils to new Positioning Folder
Refactored some of the utils so that all the related positioning utils
are put together.
