+++ src/js/TextFields/TextField.js
@@ -5,7 +5,7 @@ import deprecated from 'react-prop-types/lib/deprecated';
 import isRequiredForA11y from 'react-prop-types/lib/isRequiredForA11y';
 
 import getField from '../utils/getField';
-import getTextWidth from '../utils/getTextWidth';
+import getTextWidth from '../utils/Positioning/getTextWidth';
 import controlled from '../utils/PropTypes/controlled';
 import invalidIf from '../utils/PropTypes/invalidIf';
 import minNumber from '../utils/PropTypes/minNumber';
